+++ CHANGELOG.md
@@ -151,6 +151,7 @@
 - Presets: unified presets along different query filtering tools.
 - Presets: changed all references to rating >x to ≥X+1. i.e. 'Rating ≥3' instead of 'Rating >2'
 - UI: unified script updates settings across all my scripts, look for 'Updates' submenu.
+- UI: show shortcuts on tooltip setting moved to global setting for the toolbar.
 - Helpers: updated helpers.
 - Helpers: general code cleanup on menus internal code. Please report any bug on extra separators or menu entries not working as expected.
 ### Removed